requirements.txt
# Core dependencies for Constellation - Distributed Container Orchestration System
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.25.0
docker>=7.0.0,<8.0.0
pydantic>=2.5.0,<3.0.0
prometheus-client>=0.19.0,<1.0.0
websockets>=12.0,<13.0.0
httpx>=0.26.0
aiofiles>=23.0.0
asyncio-mqtt>=0.16.0
pyyaml>=6.0
sqlalchemy>=2.0.0
alembic>=1.13.0
redis>=5.0.0
structlog>=23.0.0
click>=8.1.0
python-multipart>=0.0.6

# Tailscale integration (using subprocess for tailscale CLI)
# Note: tailscale-client is not available, we'll use subprocess to call tailscale CLI

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0
pre-commit>=3.5.0
